- YAM_MA.c: used the modular functionality of our new internal readmail   classes to integrate a MailPreview pane similar to others found in   mail clients like Thunderbird&Co. This pane always shows the currently   selected mail which allows to read the mail without having to open any   separate window. However, as this is a new major feature the pane can   be disabled in the main YAM configuration to ensure the old behaviour   without any pane.   In addition, to make the browsing in the mail list a bit smoother, the   pane will only load the new selected mail after a specified short time   interval which can however be altered in the hidden configuration of YAM.   Per default this timeinterval is 100ms, which means that if the user   hasn't moved on in the mail list for 100ms YAM will load and display the   last selected mail into the mailpreview pane. - YAM_RE.c, ...: reworked the whole ReadWindow related codebase, separated   the window code from the readmail code and integrated it into own   internal custom MUI classes. This, aside from the more modular and   safer approach for reading mails, allows now to put an unlimited number   of readmail objects whereever required.   In fact, YAM does now support to open an unlimited number of read window   instead of just the old possible maximum of 3 simultaneous opened read   windows when the "Multiple windows" option is enabled in the YAM config.
    ATTENTION: Please note that due to the heaps of changes in the
    development of this new object-oriented readwindow facility, some
    features dependent on the old behaviour might be broken. This can
    include, but is not limited to, Arexx commands previously expecting/
    serving only for up to a maximum of 3 possible read windows. So if you
    stumble over a problem with the new readmail facility, please don't
    hestitate to bring it to our attention so that we can fix it immediatly.

+++ yamos/classes/MailTextEdit.c
@@ -202,7 +202,7 @@ OVERLOAD(MUIM_HandleEvent)
 					if(visible > 0)
 					{
 						// we scroll 1/6 of the displayed text by default
-						LONG delta = ((((float)visible)/6.0)+0.5); // round the value
+						LONG delta = ((((double)visible)/6.0)+0.5); // round the value
 
 						// make sure that we scroll at least 1 line
 						if(delta < 1) delta = 1;
